Split sizing scale helpers
